student registration remove

+++ src/application/use-cases/student/delete-student.ts
@@ -1,7 +1,9 @@
 import { Injectable } from '@nestjs/common';
 
 import { Student } from '@application/entities/student/student';
+import { CourseHistoriesRepository } from '@application/repositories/course-histories-repository';
 import { StudentsRepository } from '@application/repositories/students-repository';
+import { EntityInUse } from '../errors/entity-in-use';
 import { StudentNotFound } from '../errors/student-not-found';
 
 interface DeleteStudentResponse {
@@ -13,13 +15,19 @@ interface DeleteStudentRequest {
 
 @Injectable()
 export class DeleteStudent {
-  constructor(private studentsRepository: StudentsRepository) {}
+  constructor(private studentsRepository: StudentsRepository,
+    private courseHistoryRepository: CourseHistoriesRepository) {}
 
   async execute(request: DeleteStudentRequest): Promise<DeleteStudentResponse> {
     const student = await this.studentsRepository.findById(request.studentId);
 
     if (!student) {
       throw new StudentNotFound();
+    }
+
+    const isExistingCourseHistoryWithStudent = await this.courseHistoryRepository.findByStudent(student.studentId.toString())
+    if(isExistingCourseHistoryWithStudent.length > 0){
+      throw new EntityInUse('student')
     }
 
     await this.studentsRepository.delete(student.studentId.toString());
